--- UItesting/src/main/java/com/example/uitesting/UItestingApplication.java
@@ -1,13 +1,0 @@
-package com.example.uitesting;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UItestingApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UItestingApplication.class, args);
-    }
-
-}
