--- UItesting/src/main/java/com/example/uitesting/runner/TestRunner.java
@@ -1,12 +1,0 @@
-package com.example.uitesting.runner;
-
-import io.cucumber.testng.AbstractTestNGCucumberTests;
-import io.cucumber.testng.CucumberOptions;
-
-@CucumberOptions(
-        features = "src/test/resources/features",
-        glue = "com.example.uitesting.stepdefinitions",
-        plugin = {"pretty", "html:target/cucumber-reports.html"}
-)
-public class TestRunner extends AbstractTestNGCucumberTests {
-}
